+++ docs/changes.md
@@ -17,6 +17,9 @@ in jupyter notebooks, #802
 - in `Text2D` and `Text3D` shortcut to symbols are now formatted as `:gamma` and no more as `\gamma`
 - fix `mesh.stretch()` thanks to @mikaeltulldahl in #807
 - fixes to `cmap()`
+- added `mesh.is_manifold()` method #813
+- added `utils.open3d2vedo()` and `utils.vedo2open3d` converters as per #663
+
 
 -------------------------
 ## New/Revised Examples